app_dir {rappdirs}	R Documentation

Convenience wrapper for getting app dirs.

Description

Has methods:

Usage

app_dir(appname = NULL, appauthor = appname, version = NULL,
  expand = TRUE, os = get_os())

Arguments

appname
is the name of application. If NULL, just the system directory is returned.

appauthor
(only required and used on Windows) is the name of the appauthor or distributing body for this application. Typically it is the owning company name. This falls back to appname.

version
is an optional version path element to append to the path. You might want to use this if you want multiple versions of your app to be able to run independently. If used, this would typically be "<major>.<minor>". Only applied when appname is not NULL.

expand
If TRUE (the default) will expand the R_LIBS specifiers with their equivalents. See R_LIBS for list of all possibly specifiers.

os
Operating system whose conventions are used to construct the requested directory. Possible values are "win", "mac", "unix". If NULL (the default) then the convention of the current operating system (as determined by rappdirs:::get_os) will be used. This argument is unlikely to find much use outside package testing (see details section of user_data_dir).

Details

cache

log

data

config

site_data

site_config

Examples

ggplot2_app <- app_dir("ggplot2", "hadley")
ggplot2_app$cache()
ggplot2_app$log()
ggplot2_app$data()
ggplot2_app$config()
ggplot2_app$site_config()
ggplot2_app$site_data()

[Package rappdirs version 0.3.1 Index]