rappdirs: Application Directories: Determine Where to Save Data, Caches, and Logs

An easy way to determine which directories on the users computer you should use to save data, caches and logs. A port of Python's 'Appdirs' (<https://github.com/ActiveState/appdirs>) to R.

Main functions

user_data_dir()

user_config_dir()

user_cache_dir()

site_data_dir()

user_log_dir()

See also

Author

Maintainer: Hadley Wickham hadley@rstudio.com [translator, copyright holder]

Authors:

Sridhar Ratnakumar

Trent Mick

Trevor Davis [translator]

Other contributors:

RStudio [copyright holder]

ActiveState (R/appdir.r, R/cache.r, R/data.r, R/log.r translated from appdirs) [copyright holder]

Eddy Petrisor [contributor]

Gabor Csardi [contributor]

Gregory Jefferis [contributor]